pyproject.toml
[tool.poetry]
name = "guru-quotes"
version = "0.1.0"
description = ""
authors = ["Ivan Khokhlov <ivan.s.khokhlov@rt.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.116.1"
uvicorn = "^0.35.0"
pytest = "^8.4.1"
httpx = "^0.28.1"
pydantic = "^2.11.7"
ruff = "^0.12.8"
pydantic-settings = "^2.10.1"
isort = "^6.0.1"
email-validator = "^2.2.0"
fastapi-pagination = "^0.14.0"
python-dotenv = "^1.1.1"
psycopg2-binary = "^2.9.10"
sqlalchemy = "^2.0.43"
sqlmodel = "^0.0.24"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
